requirements.txt
PySide6>=6.6.0
SQLAlchemy>=2.0.23
reportlab>=4.0.7
bcrypt>=4.1.1
qrcode>=7.4.2
pillow>=10.1.0
alembic>=1.12.1
python-dotenv>=1.0.0
typing-extensions>=4.8.0
sqlalchemy[mypy]>=2.0.23
